Move registration form to new location Restore original authentication form
