+++ src/components/registration/Authentication.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+	<!--
+	AUTHENTICATION PLACEHOLDER: WHEN AUTHENTICATION IS ENABLED THEN CONTENTS CAN BE REPLACED WITH YOUR REAL
+	AUTHENTICATION SOLUTION CURRENTLY LOGIN WILL ALWAYS BE SUCCESSFUL.
+-->
+
+	<head>
+		<meta charset="utf-8" />
+		<title>Authentication</title>
+		<script src="../../vendor.bundle.js"></script>
+	</head>
+
+	<body class="dialog authentication">
+		<div id="Authentication-tsx" />
+		<script src="Authentication.js"></script>
+	</body>
+</html>
